fix: GroupByVariable now loads when landing an favs then Labels with a groupByValue
